review comments and more
